Move Add Tenant Configure Audit Log and Monitoring to their own tabs
