+++ portal-ui/src/screens/Console/Tenants/reducer.ts
@@ -80,8 +80,8 @@ const initialState: ITenantState = {
         exposeMinIO: true,
         exposeConsole: true,
         tenantCustom: false,
-        logSearchCustom: false,
-        prometheusCustom: false,
+        logSearchEnabled: true,
+        prometheusEnabled: true,
         logSearchVolumeSize: "5",
         logSearchSizeFactor: "Gi",
         logSearchImage: "",
@@ -632,8 +632,8 @@ export function tenantsReducer(
               exposeMinIO: true,
               exposeConsole: true,
               tenantCustom: false,
-              logSearchCustom: false,
-              prometheusCustom: false,
+              logSearchEnabled: true,
+              prometheusEnabled: true,
               logSearchVolumeSize: "5",
               logSearchSizeFactor: "Gi",
               logSearchSelectedStorageClass: "default",
